requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI/ML
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0

# Web Scraping
beautifulsoup4>=4.9.3
scrapy>=2.5.0
requests>=2.26.0

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.3

# Threat Intelligence
stix2>=3.0.0
taxii2-client>=2.3.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-asyncio>=0.15.1

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.2
loguru>=0.5.3
aiohttp>=3.8.0 